requirements.txt
attrs==23.1.0
blinker==1.6.2
blis==0.7.10
catalogue==1.0.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
cymem==2.0.7
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
Flask==2.2.5
ftfy==6.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
lxml==4.9.3
MarkupSafe==2.1.3
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.2
pandas==2.0.3
pdfminer.six==20221105
plac==1.1.3
preshed==3.0.8
pycparser==2.21
pycryptodome==3.18.0
pyresparser==1.0.6
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sortedcontainers==2.4.0
spacy==2.3.9
srsly==1.0.7
thinc==7.4.6
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
wasabi==0.10.1
wcwidth==0.2.6
Werkzeug==2.3.7
PyPDF2==3.0.1
